Fix a few restorecon calls.
restorecon /proc does not work; /proc does not support the xattr API
but is internally labeled based on genfs_contexts in policy.

fdAlbum is a file not a directory so we only need a regular restorecon,
and move it to the same location as in other device init*.rc files.

Change-Id: Iad8101039dc7abed53ff113c1cacdf9e428a47a0