Fix mode on sysfs file to avoid need for thermald dac_override.
/sys/class/leds/lcd-backlight/brightness has system owner/group
and is only owner-writable which triggers CAP_DAC_OVERRIDE check
when thermald tries to write to it.  Fix by chmod 664 to make
it writable by group system, which thermald already has in its
group set.

Before the change:
-rw-r--r-- system   system            u:object_r:sysfs:s0 brightness

After the change:
-rw-rw-r-- system   system            u:object_r:sysfs:s0 brightness

The policy for thermald is left unchanged by this change; we can
continue to watch for further dac_override granted messages until
we are sure there are no further cases.

Change-Id: I70dfb6b3de39c2fb6c45d6e8dc514238a0981797